Completely recoded HvlTiledRect.
Instead of it being in an instanced 
class, I moved it into another type of painter and is now accessible
through HvlPainter2D. It now also takes a string for the UV arguments
instead of a whole bunch of separate floats and saves some overloads,
but this is up for change (either to a new class or back to multiple
arguments). Close #138.